Resolves possible error in user variable path assumptions
This fix removes using the system `which` command for determining the
path for the following commands:
    - ls
    - grep
    - sort
    - head

It instead will use the `command -v` builtin in order to determine the
various paths to the binary applications. Use of which leaves open
the possibility for bugs and different use cases on various systems.

For more info on why this is a bad practice see:
http://stackoverflow.com/questions/592620/check-if-a-program-exists-from-a-bash-script