Fix a problem on macOS 'sed' (BSD) version
There is a required parameter for the inline edit option on macOS 'sed' version.
The original change doesn't have this and so, the inline edit option with the
missing parameter was causing macOS sed's command to fail.
This commit also double quoted the path parameters to avoid possible failures
due to usage of spaces into those pathnames.