Replaces egrep with grep -E in gvm_export_path function
`egrep` is listed as deprecated and in Fedora 38, is outputting the message "warnin: egrep is obsolescent". According to #431, the official release notes from Grep 3.8 state egrep has been deprecated since 2007.

This PR replaces `egrep` in gvm_export_path with `grep -E`.

`egrep` is mentioned in a couple of the files in `./examples/native/`, but as I don't think they're critical to the use of GVM, I have not attempted to replace them.

Fixes #431

Signed-off-by: Chris Collins <collins.christopher@gmail.com>