Fix gvm path inclusion vulnerability (#358)
If LD_LIBRARY_PATH (or DYLD_LIBRARY_PATH) is not set, the install/activate script will cause the
environment variable to be set to something like:

```
LD_LIBRARY_PATH=/home/david/.gvm/pkgsets/go1.13/global/overlay/lib:
```

This means that the dynamic library loader will attempt to load libraries from two locations:

1. "/home/david/.gvm/pkgsets/go1.13/global/overlay/lib"
2. "" which is interpreted as the current working directory

This means that it becomes unsafe to run any terminal commands from an untrusted directory. Ie,
running `ls` inside of an untrusted git repository could lead to arbitrary code execution.

In order to fix this, we simply change the shell expansion such that if LD_LIBRARY_PATH is not
set it does not include a trailing `:` at the end.